Move hepler functions for merging JSONs into a seperate module
